// src/types/open-api-schema-types.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/todos/": {
    get: operations["getTodos"];
    put: operations["updateTodo"];
    post: operations["createTodo"];
    delete: operations["deleteTodo"];
    parameters: {};
  };
}

export interface definitions {
  Todo: {
    id: number;
    message: string;
  };
}

export interface operations {
  getTodos: {
    parameters: {};
    responses: {
      /** List of todos. */
      200: {
        schema: definitions["Todo"][];
      };
    };
  };
  updateTodo: {
    parameters: {
      query: {
        id: string;
      };
      body: {
        todo?: definitions["Todo"];
      };
    };
    responses: {
      /** Updated ok */
      200: unknown;
    };
  };
  createTodo: {
    parameters: {
      body: {
        todo?: definitions["Todo"];
      };
    };
    responses: {
      /** Created */
      201: unknown;
    };
  };
  deleteTodo: {
    parameters: {
      query: {
        id: string;
      };
    };
    responses: {
      /** Delete */
      200: unknown;
    };
  };
}

export interface external {}